+++ app/services/categories/find_service.rb
@@ -1,0 +1,24 @@
+module Categories
+  class FindService < ApplicationService
+    def initialize(params)
+      @params = params
+    end
+
+    def call
+      if params[:filter].present?
+        Category.where('award_id = ? AND name ILIKE ?', params[:award_id], "%#{params[:filter]}%")
+                .paginate(page: params[:page]).order("name #{order}")
+      else
+        Category.where(award_id: params[:award_id]).paginate(page: params[:page]).order("name #{order}")
+      end
+    end
+
+    private
+
+    attr_reader :params
+
+    def order
+      params[:order].presence || 'asc'
+    end
+  end
+end
