--- app/services/category_service.rb
@@ -1,22 +1,0 @@
-class CategoryService
-  def initialize(params)
-    @params = params
-  end
-
-  def call
-    if params[:filter].present?
-      Category.where('award_id = ? AND name ILIKE ?', params[:award_id], "%#{params[:filter]}%")
-              .paginate(page: params[:page]).order("name #{order}")
-    else
-      Category.where(award_id: params[:award_id]).paginate(page: params[:page]).order("name #{order}")
-    end
-  end
-
-  private
-
-  attr_reader :params
-
-  def order
-    params[:order].presence || 'asc'
-  end
-end
