+++ app/services/movie_awards/find_service.rb
@@ -1,0 +1,19 @@
+module MovieAwards
+  class FindService < ApplicationService
+    def initialize(params)
+      @params = params
+    end
+
+    def call
+      MovieAward.where(movie_id: params[:movie_id]).paginate(page: params[:page]).order("delivery_year #{order}")
+    end
+
+    private
+
+    attr_reader :params
+
+    def order
+      params[:order].presence || 'asc'
+    end
+  end
+end
