+++ app/services/movies/find_service.rb
@@ -1,5 +1,5 @@
-module EntityManager
-  class MovieService < ApplicationService
+module Movies
+  class FindService < ApplicationService
     RELATED_ENTITIES = %i[companies_movies companies countries_movies
                           countries genres_movies genres languages_movies
                           languages movies_tags tags movie_awards comments].freeze
